Conflicts:
	src/webroot/cms/content-manager/lang.json
	src/webroot/cms/lib/supra/build/input/select.js
	src/webroot/cms/lib/supra/build/input/string.js

+++ src/lib/Symfony/Component/HttpFoundation/Tests/Session/Storage/MockFileSessionStorageTest.php
@@ -28,7 +28,7 @@ class MockFileSessionStorageTest extends \PHPUnit_Framework_TestCase
     private $sessionDir;
 
     /**
-     * @var MockFileSessionStorage
+     * @var FileMockSessionStorage
      */
     protected $storage;
 
@@ -40,14 +40,12 @@ class MockFileSessionStorageTest extends \PHPUnit_Framework_TestCase
 
     protected function tearDown()
     {
-        foreach (glob($this->sessionDir.'/mocksess_*') as $file) {
-            unlink($file);
-        }
-
-        rmdir($this->sessionDir);
-
         $this->sessionDir = null;
         $this->storage = null;
+        array_map('unlink', glob($this->sessionDir.'/*.session'));
+        if (is_dir($this->sessionDir)) {
+            rmdir($this->sessionDir);
+        }
     }
 
     public function testStart()
@@ -111,8 +109,8 @@ class MockFileSessionStorageTest extends \PHPUnit_Framework_TestCase
     private function getStorage()
     {
         $storage = new MockFileSessionStorage($this->sessionDir);
-        $storage->registerBag(new FlashBag);
-        $storage->registerBag(new AttributeBag);
+        $storage->registerBag(new FlashBag());
+        $storage->registerBag(new AttributeBag());
 
         return $storage;
     }
